backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
paho-mqtt>=1.6.0
websocket-client>=1.7.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
chromadb>=0.4.0
edge-tts>=6.1.0
